fix: change roles location
